got expo dev working again
